scripts/requirements.txt
opencv-python
pdf2image
pytesseract
Pillow